Fix attributes

+++ shaders/default-light.vert
@@ -3,7 +3,7 @@
 #extension GL_ARB_explicit_attrib_location: require
 #extension GL_ARB_explicit_uniform_location: require
 
-layout(location = 0) in ivec4 draw_info;
+layout(location = 0) in vec4 draw_info;
 layout(location = 1) in vec3 position;
 layout(location = 2) in vec3 normal;
 layout(location = 3) in vec2 texcoord;
@@ -26,7 +26,7 @@ layout(location = 0) flat out int draw_id;
 
 void main()
 {
-    draw_id = draw_info[0];
+    draw_id = int(draw_info[0]);
     mat4 transform = transforms[draw_id];
     gl_Position = view_projection * transform * vec4(position, 1.0);
 }
